requirements.txt
2captcha-python==1.2.1
Django==4.2.6
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.3
djangorestframework==3.14.0
Markdown==3.5
Pillow==10.0.1
PyAutoGUI==0.9.54
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.13.0
undetected-chromedriver==3.5.3
webdriver-manager==4.0.1
